+++ src/neon/azext_neon/aaz/latest/neon/postgres/organization/_delete.py
@@ -17,7 +17,7 @@ from azure.cli.core.aaz import *
     confirmation="Are you sure you want to perform this operation?",
 )
 class Delete(AAZCommand):
-    """Delete a Neon Resource
+    """Delete a Neon organization
 
     :example: Organizations_Delete
         az neon postgres organization delete --subscription 12345678-1234-1234-1234-123456789abc --resource-group demoResourceGroup --name demoNeonResource
@@ -49,7 +49,7 @@ class Delete(AAZCommand):
         _args_schema = cls._args_schema
         _args_schema.name = AAZStrArg(
             options=["-n", "--name"],
-            help="Name of the Neon resource",
+            help="Name of the Neon organization",
             required=True,
             id_part="name",
             fmt=AAZStrArgFormat(
@@ -59,7 +59,7 @@ class Delete(AAZCommand):
             ),
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
-            help="Name of the Resource Group",
+            help="Name of the resource group",
             required=True,
         )
         return cls._args_schema